# crypto/merkle: optimize merkle tree hashing