--- a/CHANGELOG_PENDING.md
+++ b/CHANGELOG_PENDING.md
@@ -99,7 +99,7 @@ Friendly reminder: We have a [bug bounty program](https://hackerone.com/tendermi
 - [statesync] \#6378 Retry requests for snapshots and add a minimum discovery time (5s) for new snapshots.
 - [node/state] \#6370 graceful shutdown in the consensus reactor (@JayT106)
 - [crypto/merkle] \#6443 Improve HashAlternatives performance (@cuonglm)
-- [crypto/merkle] \#6513 Optmize HashAlternatives (@marbar3778)
+- [crypto/merkle] \#6513 Optimize HashAlternatives (@marbar3778)
 
 ### BUG FIXES